# Conflicts:
#	beacon-chain/blockchain/head_sync_committee_info.go
#	beacon-chain/blockchain/process_attestation_helpers.go
#	beacon-chain/blockchain/process_block_test.go
#	beacon-chain/blockchain/service.go
#	beacon-chain/core/altair/upgrade.go
#	beacon-chain/core/execution/upgrade.go
#	beacon-chain/core/execution/upgrade_test.go
#	beacon-chain/core/transition/transition.go
#	beacon-chain/rpc/eth/validator/validator.go
#	beacon-chain/slasher/detect_blocks_test.go
#	beacon-chain/slasher/process_slashings_test.go
#	beacon-chain/state/state-native/v1/BUILD.bazel
#	beacon-chain/state/state-native/v1/beacon_state_mainnet.go
#	beacon-chain/state/state-native/v1/generated.ssz.go
#	beacon-chain/state/state-native/v1/getters_attestation.go
#	beacon-chain/state/state-native/v1/getters_block.go
#	beacon-chain/state/state-native/v1/getters_block_test.go
#	beacon-chain/state/state-native/v1/getters_checkpoint.go
#	beacon-chain/state/state-native/v1/getters_eth1.go
#	beacon-chain/state/state-native/v1/getters_misc.go
#	beacon-chain/state/state-native/v1/getters_randao.go
#	beacon-chain/state/state-native/v1/getters_state.go
#	beacon-chain/state/state-native/v1/getters_test.go
#	beacon-chain/state/state-native/v1/getters_validator.go
#	beacon-chain/state/state-native/v1/setters_attestation.go
#	beacon-chain/state/state-native/v1/setters_attestation_test.go
#	beacon-chain/state/state-native/v1/setters_block.go
#	beacon-chain/state/state-native/v1/setters_misc.go
#	beacon-chain/state/state-native/v1/setters_randao.go
#	beacon-chain/state/state-native/v1/setters_state.go
#	beacon-chain/state/state-native/v1/setters_validator.go
#	beacon-chain/state/state-native/v1/state_trie.go
#	beacon-chain/state/state-native/v1/types.go
#	beacon-chain/state/state-native/v2/BUILD.bazel
#	beacon-chain/state/state-native/v2/getters_block.go
#	beacon-chain/state/state-native/v2/getters_block_test.go
#	beacon-chain/state/state-native/v2/getters_checkpoint.go
#	beacon-chain/state/state-native/v2/getters_eth1.go
#	beacon-chain/state/state-native/v2/getters_misc.go
#	beacon-chain/state/state-native/v2/getters_participation.go
#	beacon-chain/state/state-native/v2/getters_randao.go
#	beacon-chain/state/state-native/v2/getters_state.go
#	beacon-chain/state/state-native/v2/getters_sync_committee.go
#	beacon-chain/state/state-native/v2/getters_test.go
#	beacon-chain/state/state-native/v2/getters_validator.go
#	beacon-chain/state/state-native/v2/setters_block.go
#	beacon-chain/state/state-native/v2/setters_misc.go
#	beacon-chain/state/state-native/v2/setters_randao.go
#	beacon-chain/state/state-native/v2/setters_state.go
#	beacon-chain/state/state-native/v2/setters_validator.go
#	beacon-chain/state/state-native/v2/state_trie.go
#	beacon-chain/state/state-native/v2/state_trie_test.go
#	beacon-chain/state/state-native/v2/types.go
#	beacon-chain/state/state-native/v3/BUILD.bazel
#	beacon-chain/state/state-native/v3/getters_block.go
#	beacon-chain/state/state-native/v3/getters_block_test.go
#	beacon-chain/state/state-native/v3/getters_checkpoint.go
#	beacon-chain/state/state-native/v3/getters_eth1.go
#	beacon-chain/state/state-native/v3/getters_misc.go
#	beacon-chain/state/state-native/v3/getters_participation.go
#	beacon-chain/state/state-native/v3/getters_payload_header.go
#	beacon-chain/state/state-native/v3/getters_randao.go
#	beacon-chain/state/state-native/v3/getters_state.go
#	beacon-chain/state/state-native/v3/getters_sync_committee.go
#	beacon-chain/state/state-native/v3/getters_test.go
#	beacon-chain/state/state-native/v3/getters_validator.go
#	beacon-chain/state/state-native/v3/setters_block.go
#	beacon-chain/state/state-native/v3/setters_misc.go
#	beacon-chain/state/state-native/v3/setters_randao.go
#	beacon-chain/state/state-native/v3/setters_state.go
#	beacon-chain/state/state-native/v3/setters_validator.go
#	beacon-chain/state/state-native/v3/state_trie.go
#	beacon-chain/state/state-native/v3/state_trie_test.go
#	beacon-chain/state/state-native/v3/types.go
#	hack/update-go-ssz.sh
#	proto/eth/v2/BUILD.bazel
#	proto/prysm/v1alpha1/BUILD.bazel
#	testing/fuzz/BUILD.bazel
#	testing/fuzz/block_fuzz.go
#	testing/fuzz/rpc_status_fuzz.go
#	testing/fuzz/state_fuzz.go
#	testing/spectest/shared/bellatrix/fork/upgrade_to_altair.go

+++ beacon-chain/sync/initial-sync/service_test.go
@@ -27,7 +27,7 @@ import (
 )
 
 func TestService_Constants(t *testing.T) {
-	if uint64(params.BeaconConfig().MaxPeersToSync)*flags.Get().BlockBatchLimit > uint64(1000) {
+	if params.BeaconConfig().MaxPeersToSync*flags.Get().BlockBatchLimit > 1000 {
 		t.Fatal("rpc rejects requests over 1000 range slots")
 	}
 }
